Deployment/requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
streamlit
opencv-contrib-python-headless
ultralytics
onnxruntime
scikit-learn==1.6.1
plotly